requirements.txt
Pillow
manga-ocr
easyocr
opencv-python
googletrans
pyphen
iopaint
torch
numpy